// types.ts
export interface CardData {
  companyNameMain: string; // Dimo
  companyNameSuffix: string; // CONSTRUCTION
  companyType: string; // EOOD
  slogan: string; // ПРОЕКТИРАНЕ НА СГРАДИ И СЪОРЪЖЕНИЯ
  address: string;
  email1: string;
  email2: string;
  website: string;
  phone1: string;
  person1: string;
  phone2: string;
  person2: string;
  phone3: string;
  person3: string;
}

export const INITIAL_CARD_DATA: CardData = {
  companyNameMain: "Dimo",
  companyNameSuffix: "CONSTRUCTION",
  companyType: "OOD",
  slogan: "ПРОЕКТИРАНЕ НА СГРАДИ И СЪОРЪЖЕНИЯ",
  address: "офис: гр.Стара Загора,бул.“Цар Симеон Велики“ №4,ет.5,ап.10",
  email1: "dimovconstruction.sz@gmail.com",
  email2: "",
  website: "www.dimovconstruction.com",
  phone1: "0888 536401",
  person1: "инж.Пламен Димов",
  phone2: "0883 386003",
  person2: "инж.Светлозар Димов",
  phone3: "0888 574164",
  person3: "инж.Димо Димов"
};